[CoreBundle] Moved all models to the same directory
