+++ src/TwoMartens/Bundle/CoreBundle/Tests/Model/OptionTest.php
@@ -7,9 +7,9 @@
  * file that was distributed with this source code.
  */
 
-namespace TwoMartens\Bundle\CoreBundle\Tests\Model\Option;
+namespace TwoMartens\Bundle\CoreBundle\Tests\Model;
 
-use TwoMartens\Bundle\CoreBundle\Model\Option\Option;
+use TwoMartens\Bundle\CoreBundle\Model\Option;
 
 /**
  * Tests the Option class.
